@@ -11,16 +11,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "13.2.0",
-    "@angular/cdk": "13.2.0",
-    "@angular/common": "13.2.0",
-    "@angular/compiler": "13.2.0",
-    "@angular/core": "13.2.0",
-    "@angular/forms": "13.2.0",
-    "@angular/material": "13.2.0",
-    "@angular/platform-browser": "13.2.0",
-    "@angular/platform-browser-dynamic": "13.2.0",
-    "@angular/router": "13.2.0",
+    "@angular/animations": "14.3.0",
+    "@angular/cdk": "^14.2.7",
+    "@angular/common": "14.3.0",
+    "@angular/compiler": "14.3.0",
+    "@angular/core": "14.3.0",
+    "@angular/forms": "14.3.0",
+    "@angular/material": "^14.2.7",
+    "@angular/platform-browser": "14.3.0",
+    "@angular/platform-browser-dynamic": "14.3.0",
+    "@angular/router": "14.3.0",
     "@ngneat/until-destroy": "^9.2.1",
     "ngx-toastr": "^14.3.0",
     "rxjs": "~7.4.0",
@@ -28,19 +28,19 @@
     "zone.js": "0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "13.2.0",
-    "@angular-eslint/builder": "13.5.0",
-    "@angular-eslint/eslint-plugin": "13.5.0",
-    "@angular-eslint/eslint-plugin-template": "13.5.0",
-    "@angular-eslint/schematics": "13.5.0",
-    "@angular-eslint/template-parser": "13.5.0",
-    "@angular/cli": "13.2.0",
-    "@angular/compiler-cli": "13.2.0",
+    "@angular-devkit/build-angular": "14.2.13",
+    "@angular-eslint/builder": "^14.4.0",
+    "@angular-eslint/eslint-plugin": "^14.4.0",
+    "@angular-eslint/eslint-plugin-template": "^14.4.0",
+    "@angular-eslint/schematics": "^14.4.0",
+    "@angular-eslint/template-parser": "^14.4.0",
+    "@angular/cli": "14.2.13",
+    "@angular/compiler-cli": "14.3.0",
     "@types/node": "^17.0.12",
     "@typescript-eslint/eslint-plugin": "^5.27.1",
     "@typescript-eslint/parser": "^5.27.1",
     "eslint": "^8.17.0",
     "eslint-plugin-rxjs": "^5.0.2",
-    "typescript": "~4.4.3"
+    "typescript": "~4.8.4"
   }
 }
